Fixes: Issue #10
Problem: Call to BigIP resource in virtual server does not use proper
named parameters

Analysis: Change the create call in the virtual server plugin to use the
correct properties for each named parameter.

Unit/Robot tests:
