+++ f5_heat/examples/composite.yaml
@@ -1,0 +1,188 @@
+# Copyright 2015-2016 F5 Networks Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+heat_template_version: 2015-04-30
+
+description: Template using all F5 resource types
+
+parameters:
+  ve_instance1:
+    type: string
+    label: IP of First VE Instance
+    default: 10.190.6.7
+  ve_instance2:
+    type: string
+    label: IP of Second VE Instance
+    default: 10.190.6.4
+  bigip_un:
+    type: string
+    label: Username of BigIP Login
+    default: admin
+  bigip_pw:
+    type: string
+    label: Password of BigIP Login
+    default: admin
+  vs_name:
+    type: string
+    label: Name of Virtual Server
+    default: vs_resource
+  vs_ip:
+    type: string
+    label: IP of Virtual Server
+    default: 140.0.0.1
+  vs_port:
+    type: string
+    label: Port of Virtual Server
+    default: 80
+  pool1_name:
+    type: string
+    label: Name of Pool 1
+    default: pool_resource
+  pool2_name:
+    type: string
+    label: Name of Pool 2
+    default: pool_resource2
+  iapp_template_name:
+    type: string
+    label: Name of iApp Template
+    default: iapp_template_resource
+  iapp_service_name:
+    type: string
+    label: Name of iApp Service
+    default: iapp_service_resource
+
+parameter_groups:
+  - label: VE Parameters
+    parameters:
+      - ve_instance1
+      - ve_instance2
+      - bigip_un
+      - bigip_pw
+  - label: Virtual Server Settings
+    parameters:
+      - vs_name
+      - vs_ip
+      - vs_port
+  - label: Pool Settings
+    parameters:
+      - pool1_name
+      - pool2_name
+  - label: iApp Settings
+    parameters:
+      - iapp_template_name
+      - iapp_service_name
+
+resources:
+  bigip1:
+    type: F5::BigIP::Device
+    properties:
+      ip: { get_param: ve_instance1 }
+      username: { get_param: bigip_un }
+      password: { get_param: bigip_pw }
+  bigip2:
+    type: F5::BigIP::Device
+    properties:
+      ip: { get_param: ve_instance2 }
+      username: { get_param: bigip_un }
+      password: { get_param: bigip_pw }
+  virtual_server:
+    type: F5::LTM::VirtualServer
+    properties:
+      bigip_server: { get_resource: bigip1 }
+      name: { get_param: vs_name }
+      ip: { get_param: vs_ip }
+      port: { get_param: vs_port }   
+  pool1:
+    type: F5::LTM::Pool
+    properties:
+      bigip_server: { get_resource: bigip1 }
+      name: { get_param: pool1_name }
+      members:
+        - member_ip: 140.0.0.3
+          member_port: 80
+        - member_ip: 140.0.0.4
+          member_port: 80
+  pool2:
+    type: F5::LTM::Pool
+    depends_on: bigip2
+    properties:
+      bigip_server: { get_resource: bigip2 }
+      name: { get_param: pool2_name }
+      members:
+        - member_ip: 150.0.0.1
+          member_port: 8080
+  iapp_service:
+    type: F5::Sys::iAppService
+    depends_on: iapp_template
+    properties:
+      name: { get_param: iapp_service_name }
+      bigip_server: { get_resource: bigip1 }
+      template_name: { get_param: iapp_template_name }
+  iapp_template:
+    type: F5::Sys::iAppTemplate
+    depends_on: bigip1
+    properties:
+      name: { get_param: iapp_template_name }
+      bigip_server: { get_resource: bigip1 }
+      requires_modules: [ ltm ]
+      implementation: |
+        #TMSH-VERSION: 11.6.0
+        iapp::template start
+        tmsh::create {
+          ltm pool http_pool
+          description "A pool of http servers"
+          load-balancing-mode least-connections-node
+          members replace-all-with {
+              129.0.0.1:8000 {
+                  address 129.0.0.1
+              }
+              130.0.0.1:8000 {
+                  address 130.0.0.1
+              }
+              140.0.0.1:8000 {
+                  address 140.0.0.1
+              }
+              150.0.0.1:8000 {
+                  address 150.0.0.1
+              }
+          }
+        }
+        tmsh::create {
+           ltm virtual http_vs
+           destination 10.15.15.30:80
+           ip-protocol tcp
+           mask 255.255.255.255
+           pool http_pool
+           profiles replace-all-with {
+               http { }
+               tcp { }
+           }
+           source 0.0.0.0/0
+           translate-address enabled
+           translate-port enabled
+        }
+        tmsh::create {
+           ltm virtual-address 10.4.4.115
+           address 10.4.4.115
+           arp enabled
+           icmp-echo enabled
+           mask 255.255.255.255
+           traffic-group traffic-group-1
+        }
+        iapp::template stop
+      presentation: |
+        section say_hello {
+          message intro "This template deploys a virtual server and a pool with several members."
+        }
