+++ f5_heat/resources/f5_bigip_device.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 #
 
-from f5.bigip.bigip import BigIP
+from f5.bigip import BigIP
 from heat.common.i18n import _
 from heat.engine import properties
 from heat.engine import resource
