+++ f5_heat/resources/test/test_f5_bigip_device.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 #
 
-from f5.bigip.bigip import BigIP
+from f5.bigip import BigIP
 from f5_heat.resources import f5_bigip_device
 from heat.common import exception
 from heat.common import template_format
@@ -23,6 +23,7 @@ from heat.engine import template
 
 import mock
 import pytest
+import uuid
 
 f5_bigip_defn = '''
 heat_template_version: 2015-04-30
@@ -51,19 +52,11 @@ resources:
 '''
 
 
-test_uuid = '5abe95ca-0bc9-4158-b51b-366156ea9448'
-
-versions = ('2015-04-30', '2015-04-30')
-
-
-@mock.patch.object(template, 'get_version', return_value=versions)
-@mock.patch.object(
-    template,
-    'get_template_class',
-    return_value=HOTemplate20150430
-)
-def mock_template(templ_vers, templ_class, test_templ=f5_bigip_defn):
+def mock_template(test_templ=f5_bigip_defn):
     '''Mock a Heat template for the Kilo version.'''
+    versions = ('2015-04-30', '2015-04-30')
+    template.get_version = mock.Mock(return_value=versions)
+    template.get_template_class = mock.Mock(return_value=HOTemplate20150430)
     templ_dict = template_format.parse(test_templ)
     return templ_dict
 
@@ -86,7 +79,6 @@ def F5BigIP():
     f5_bigip_obj = f5_bigip_device.F5BigIPDevice(
         "testing_service", rsrc_def, mock.MagicMock()
     )
-    f5_bigip_obj.uuid = test_uuid
     f5_bigip_obj.validate()
     return f5_bigip_obj
 
@@ -111,16 +103,15 @@ def test__init__error(mocked_bigip):
 
 def test_handle_create(F5BigIP):
     # Set uuid on resource object because stack is mocked out
+    F5BigIP.uuid = uuid.uuid4()
     create_result = F5BigIP.handle_create()
     assert create_result is None
     assert F5BigIP.resource_id is not None
 
 
 def test_handle_delete(F5BigIP):
-    assert hasattr(F5BigIP, 'bigip')
     delete_result = F5BigIP.handle_delete()
     assert delete_result is None
-    assert not hasattr(F5BigIP, 'bigip')
 
 
 def test_bigip_getter(F5BigIP):
@@ -129,7 +120,7 @@ def test_bigip_getter(F5BigIP):
 
 
 def test_bad_property():
-    template_dict = mock_template(test_templ=bad_f5_bigip_defn)
+    template_dict = mock_template(bad_f5_bigip_defn)
     rsrc_def = create_resource_definition(template_dict)
     f5_bigip_obj = f5_bigip_device.F5BigIPDevice(
         'test',
